Fix issue where remote notifications were being treated as silent when application is in the foreground
Summary:
Remote notifications that are non-silent were being handled as silent notifications when the application is in the foreground (at least on iOS 10.0.1).

Also addresses an iOS 10.0.1 issue where tapping on a remote notification while the application is in the background would not be handled by the application.

Version: Patch

Test Plan: iOS app integration / Unit Tests

Reviewers: eric

Reviewed By: eric

Subscribers: jenkins

Differential Revision: https://phabricator.automatic.co/D17879